Renamed ambiguousgeneagent.ts to determineAmbiguousGene.ts
