+++ server/routes/termdb.chat2.ts
@@ -4,7 +4,7 @@ import { classifyQuery } from './chat/classify1.ts'
 import { classifyNotPlot } from './chat/classify2.ts'
 import { classifyPlotType } from './chat/plot.ts'
 import { extract_DE_search_terms_from_query } from './chat/DEagent.ts'
-import { determineAmbiguousGenePrompt } from './chat/ambiguousgeneagent.ts'
+import { determineAmbiguousGenePrompt } from './chat/determineAmbiguousGene.ts'
 import { extract_summary_terms } from './chat/summaryagent.ts'
 import { extract_matrix_search_terms_from_query } from './chat/matrixagent.ts'
 import { extract_samplescatter_terms_from_query } from './chat/samplescatteragent.ts'
